# Jorje.TheWorld.Api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "Trips API"
  },
  "paths": {
    "/api/account": {
      "post": {
        "tags": [
          "Account"
        ],
        "operationId": "ApiAccountPost",
        "consumes": [
          "application/json",
          "text/json",
          "application/json-patch+json",
          "application/xml",
          "text/xml"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "Credentials",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Credentials"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/account/sign-in": {
      "post": {
        "tags": [
          "Account"
        ],
        "operationId": "ApiAccountSign-inPost",
        "consumes": [
          "application/json",
          "text/json",
          "application/json-patch+json",
          "application/xml",
          "text/xml"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "Credentials",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Credentials"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/persons/{id}": {
      "get": {
        "tags": [
          "Persons"
        ],
        "operationId": "ApiPersonsByIdGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Persons"
        ],
        "operationId": "ApiPersonsByIdPut",
        "consumes": [
          "application/json",
          "text/json",
          "application/json-patch+json",
          "application/xml",
          "text/xml"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "personInput",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/PersonDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Persons"
        ],
        "operationId": "ApiPersonsByIdDelete",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "patch": {
        "tags": [
          "Persons"
        ],
        "operationId": "ApiPersonsByIdPatch",
        "consumes": [
          "application/json",
          "text/json",
          "application/json-patch+json",
          "application/xml",
          "text/xml"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "patchDoc",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/JsonPatchDocument[PersonDTO]"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/persons": {
      "post": {
        "tags": [
          "Persons"
        ],
        "operationId": "ApiPersonsPost",
        "consumes": [
          "application/json",
          "text/json",
          "application/json-patch+json",
          "application/xml",
          "text/xml"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "personIput",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/PersonDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Credentials": {
      "required": [
        "email",
        "password"
      ],
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "maxLength": 100,
          "minLength": 6,
          "type": "string"
        }
      }
    },
    "PersonDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "dateOfBirth": {
          "format": "date-time",
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "personImagePath": {
          "type": "string"
        }
      }
    },
    "JsonPatchDocument[PersonDTO]": {
      "type": "object",
      "properties": {
        "operations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Operation[PersonDTO]"
          },
          "readOnly": true
        }
      }
    },
    "Operation[PersonDTO]": {
      "type": "object",
      "properties": {
        "value": {
          "type": "object"
        },
        "path": {
          "type": "string"
        },
        "op": {
          "type": "string"
        },
        "from": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}